Implementação do Novo Design
