+++ client/src/components/Chat/Chat.jsx
@@ -13,9 +13,9 @@ export default function Chat() {
 
     const matches = useMediaQuery('(min-width: 600px)')
 
-    const [text, setText] = useState('')
+    const [ text, setText ] = useState('')
 
-    const [showEmojiPicker, setShowEmojiPicker] = useState(false)
+    const [ showEmojiPicker, setShowEmojiPicker ] = useState(false)
 
     const onEmojiClick = (emojiObject) => {
         setText(prevInput => prevInput + emojiObject.emoji)
