+++ client/src/components/index.js
@@ -1,6 +1,6 @@
-import DiscoverSection from './Home/DiscoverSection'
-import AboutSection from './Home/AboutSection'
-import ComunitySection from './Home/ComunitySection'
+import DiscoverSection from './LandingPage/DiscoverSection'
+import AboutSection from './LandingPage/AboutSection'
+import ComunitySection from './LandingPage/ComunitySection'
 import Chat from './Chat/Chat'
 import Notice from './News/LittleNotice'
 import SideBar from './SideBar/SideBar'
@@ -8,7 +8,7 @@ import Folder from './Materials/Folder'
 import File from './Materials/File'
 import Contact from './Chat/Contact'
 import Classes from './Classes/Classes'
-import Post from './Timeline/Post'
+import Post from './Home/Post'
 import Comment from './Timeline/Comment'
 import Login from './Auth/Login'
 import Register from './Auth/Register'
