+++ client/src/layout/CustomInput.jsx
@@ -1,0 +1,45 @@
+import { Avatar, Box, TextField } from '@mui/material'
+
+export default function CustomInput({ inputWidth, width, icon, bgcolor, color, sx, value, defaultValue, placeholder, onChange, fullWidth, iconColor }) {
+    return (
+        <Box width={width || '110%'} display='flex'>
+            <TextField
+                width={inputWidth}
+                placeholder={placeholder}
+                variant="outlined"
+                fullWidth={fullWidth}
+                value={value}
+                defaultValue={defaultValue}
+                onChange={onChange}
+                sx={{ 
+                    backgroundColor: bgcolor, 
+                    width: '100%',
+                    height: '3.5rem',
+                    borderRadius: '15px',
+                    '& .MuiOutlinedInput-root': {
+                        '& > fieldset': {
+                            border: 'none'
+                        }
+                    },
+
+                    ...sx
+                }}
+            />
+            <Avatar  variant="rounded" sx={{ 
+                position: 'relative',
+                right: '3.5rem',
+                top: '0.5rem',
+                fontSize: '0.75rem',
+                height: '2rem',
+                width: '2rem',
+                borderRadius: '15px',
+                backgroundColor: color,
+                color: iconColor || 'text.primary',
+                border: '1px solid rgba(0, 0, 0, 0.15)',
+                p: 2.5 
+            }}>
+                {icon}
+            </Avatar>
+        </Box>
+    )
+}
