+++ client/src/layout/Footer.jsx
@@ -8,18 +8,15 @@ import FlagDE from '../assets/svg/Flags/DE.svg'
 import FlagFR from '../assets/svg/Flags/FR.svg'
 import FlagES from '../assets/svg/Flags/ES.svg'
 import FlagRU from '../assets/svg/Flags/RU.svg'
-import FlagAR from '../assets/svg/Flags/AR.svg'
+import FlagSA from '../assets/svg/Flags/SA.svg'
 import FlagCN from '../assets/svg/Flags/CN.svg'
 import FlagJP from '../assets/svg/Flags/JP.svg'
 import FlagIN from '../assets/svg/Flags/IN.svg'
 
-
-import { useNavigate } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
 
 export default function Footer() {
-    const navigate = useNavigate()
-
+    const lang = navigator.language.split('-')[1].toLowerCase()
     const { t } = useTranslation()
 
     return (
@@ -59,10 +56,9 @@ export default function Footer() {
                     <Box>
                         <Box>
                             <Select onChange={e => {
-                                navigate(`/home/${e.target.value}`)
-                                window.location.reload()
                                 localStorage.setItem('lang', e.target.value)
-                            }} fullWidth sx={{ mt: 5, pl: 4, height: 50 }} defaultValue='br' value={localStorage.getItem('lang')}>
+                                window.location.href = `/${e.target.value}`
+                            }} fullWidth sx={{ mt: 5, pl: 4, height: 50 }} defaultValue={lang || 'br'} value={localStorage.getItem('lang')}>
                                 <MenuItem value='br' >
                                     <Box display='flex' position='relative' bottom={3}>
                                         <img width='30px' style={{ marginRight: 10, position: 'relative', top: 5 }} height='30px' src={FlagBR}/>
@@ -113,7 +109,7 @@ export default function Footer() {
                                 </MenuItem>
                                 <MenuItem value='ar' >
                                     <Box display='flex' position='relative' bottom={3}>
-                                        <img width='30px' style={{ marginRight: 10, position: 'relative', top: 5 }} height='30px' src={FlagAR}/>
+                                        <img width='30px' style={{ marginRight: 10, position: 'relative', top: 5 }} height='30px' src={FlagSA}/>
                                         <Typography position='relative' top={8}>
                                             AR
                                         </Typography>
