+++ client/src/layout/SideBar.jsx
@@ -1,0 +1,9 @@
+import { Box } from '@mui/material'
+
+export default function SideBar() {
+    return (
+        <Box height='100%' width='15%'>
+
+        </Box>
+    )
+}
