+++ client/src/lib/redux/reducers/userReducer.js
@@ -2,13 +2,13 @@ import { ActionTypes } from '../../../constants'
 
 const { FETCH_ALL, CREATE, LOGOUT, LOGIN } = ActionTypes
 
-export default (state=[], action) => {
+export default (state = [], action) => {
     switch (action.type) {
     case FETCH_ALL:
         return action.payload
             
     case CREATE:
-        return [...state, action.payload]
+        return [ ...state, action.payload ]
 
     case LOGIN:
         return action.payload
@@ -16,6 +16,6 @@ export default (state=[], action) => {
     case LOGOUT:
         return action.payload
     default:
-        return [...state]
+        return [ ...state ]
     }
 }
