+++ client/src/main.jsx
@@ -5,7 +5,12 @@ import reducers from './lib/redux/reducers'
 import { legacy_createStore, applyMiddleware, compose } from 'redux'
 import { Provider } from 'react-redux'
 
-const store = legacy_createStore(reducers, compose(applyMiddleware(thunk)))
+// window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
+
+const store = legacy_createStore(
+    reducers,
+    compose(applyMiddleware(thunk)),
+)
 
 ReactDOM.createRoot(document.getElementById('root')).render(
     <Provider store={store}>
