+++ client/src/pages/index.js
@@ -1,12 +1,8 @@
-import Home from './Home'
+import LandingPage from './LandingPage'
 import ErrorPage from './ErrorPage'
-import ChatPage from './ChatPage'
-import ClassesPage from './ClassesPage'
 import FavoritesPage from './FavoritesPage'
-import MaterialsPage from './MaterialsPage'
-import NewsPage from './NewsPage'
-import TimelinePage from './TimelinePage'
 import Clips from '../components/Timeline/Clips'
+import Home from './Home'
 import Suggestion from '../components/Timeline/Suggestion'
 
-export { Home, ErrorPage, ChatPage, ClassesPage, FavoritesPage, MaterialsPage, NewsPage, TimelinePage, Clips, Suggestion }
+export { LandingPage, ErrorPage, FavoritesPage, Clips, Suggestion, Home }
